requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb 
tantivy
streamlit 
pandas
yfinance
openai 
duckduckgo-search 
yfinance 
sqlalchemy 
fastapi[standard]
groq
sentence-transformers
numpy
agentql
ipykernel
ipywidgets
httpx
ollama
langgraph
langgraph-cli[inmem]
pydantic
langchain_community
langchain_core
langchain_groq
googlesearch-python 
pycountry
pygithub